+++ crates/sdk/src/builders/mod.rs
@@ -21,6 +21,9 @@ pub mod market;
 /// Instruction builders related to market state.
 pub mod market_state;
 
+/// Instruction builders related to deposit.
+pub mod deposit;
+
 /// Instruction builders related to liquidity provider program.
 #[cfg(liquidity_provider)]
 pub mod liquidity_provider;
@@ -33,4 +36,8 @@ pub mod callback;
 /// Nonce Bytes.
 pub type NonceBytes = StringPubkey;
 
-pub use self::store_program::StoreProgram;
+pub use self::{
+    market::MarketTokenIxBuilder,
+    store_program::{StoreProgram, StoreProgramIxBuilder},
+    utils::PoolTokenHint,
+};